% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Title}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{dataframe}

\item{varx}{x variable name in df}

\item{vary}{y variable name in df}
}
\value{
ggplot scatterplot of vary vs varx
}
\description{
Title
}
\examples{
df <- data.frame(x = rnorm(10), y = rnorm(10))
diamond_plot(df, 'x', 'y')
}
