% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fccp_closed.R
\name{fccp_closed0}
\alias{fccp_closed0}
\alias{fccp_closed}
\title{Cohort Component Population Projection Model based on a Closed Female Population.}
\usage{
fccp_closed0(n = NULL, x = df0$x, p = NULL, Nx = NULL, sx = NULL,
  fx = NULL, sn = NULL, sex_ratio = 1/(1 + 1.05), tidy_output = TRUE,
  age_lab = x, gender_lab = "Female", ...)

fccp_closed(n = NULL, x = df0$x, p = NULL, Nx = NULL, sx = NULL,
  fx = NULL, sn = NULL, sex_ratio = 1/(1 + 1.05), tidy_output = TRUE,
  age_lab = x, gender_lab = "Female", ...)
}
\arguments{
\item{n}{Numeric value for the number of projection steps.}

\item{x}{Vector containing a character string of age group labels.}

\item{p}{Numeric value for step size of the population projection.}

\item{Nx}{Vector containing numeric values of the initial female population size in each age group (\code{x}).}

\item{sx, fx}{Vector containing numeric values of the age specific female survival and fertility rates.

If \code{fccp_closed0} is used a single vector is required.

If \code{fccp_closed} is used a matrix of period specific rates is required, where rows represent age groups and columns future periods. If a single vector is be passed to \code{fccp_closed} a matrix based on constant assumptions in all future rates will be constructed.}

\item{sn, sex_ratio}{Numeric value of the survivorship of newborn female babies from birth to the end of the interval and the sex ratio at birth of newborn babies.

If \code{fccp_closed0} is used a single value is required.

If \code{fccp_closed} is used a vector of period specific values is required. If a single value is be passed to \code{fccp_closed} a vector based on constant assumptions in all future rates will be constructed.}

\item{tidy_output}{Logical value to indicate if projection output should be in a tidy data format (\code{TRUE}, the default) or as a \code{matrix} where rows represent age and gender groups and columns the projection year.}

\item{age_lab, gender_lab}{Vector containing a character string of age and gender group labels. Only used if projection output is in a tidy data format. See \code{tidy_pp} for more details.}

\item{...}{Additional arguments passed to \code{\link{tidy_pp}} to build a tidy data frame (if chosen from \code{tidy_output}).}
}
\value{
Projected populations by age and gender for \code{n} future steps, given the age specific fertility and survivorship rates. Depending on the \code{tidy_output} value the projections will be returned as either a matrix or a tibble. Both versions contain the initial population sizes given in \code{Nx}.

\code{fccp_closed0} produces population projections based stricly on constant future rates.

\code{fccp_closed} produces population projections based non-constant future rates.
}
\description{
A cohort component projection model based on a closed female population,
\deqn{ \mathbf N(t+5) =  \mathbf L \mathbf N(t)  }
where the Leslie matrix, eqn{\mathbf L}, is created given user defined age specific fertility and survivorship rates.
}
\examples{
df0 <- sweden1993

# matrix output
fccp_closed0(n = 5, x = df0$x, p = 5, Nx = df0$Nx_f,
             sx = df0$sx_f, fx = df0$fx, sn = df0$Lx_f[1]/(5*100000),
             tidy_output = FALSE)

# tidy data frame output
fccp_closed0(n = 5, x = df0$x, p = 5, Nx = df0$Nx_f,
             sx = df0$sx_f, fx = df0$fx, sn = df0$Lx_f[1]/(5*100000),
             year0 = 1993, age_lab = df0$age)

# setting up non-constant future age specific fertility rates
ff <- matrix(df0$fx, nrow = length(df0$fx), ncol = 5)
ff <- sweep(ff, 2, seq(from = 1, to = 1.5, length = 5), "*")
# tfr increase
5 * colSums(ff)
# run projection with increasing fx, sx remains constant
fccp_closed(n = 5, x = df0$x, p = 5, Nx = df0$Nx_f,
            sx = df0$sx_f, fx = ff, sn = df0$Lx_f[1]/(5*100000),
            year0 = 1993, age_lab = df0$age,
            tidy_output = FALSE)
}
