% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{Population Pyramid Using ggplot2}
\usage{
pyramid(data = NULL, m = NULL, gender_lab = NULL, age_col = "Age",
  gender_col = "Gender", pop_col = "Population")
}
\arguments{
\item{data}{data frame containing columns on age, gender and population size.}

\item{m}{Upper limit on the horizontal axis}

\item{gender_lab}{Vector of length two containing a character string of gender group labels in \code{data}. Female label first, male second. If left as \code{NULL}, the function will guess labels from the unique values in the \code{gender_col}.}

\item{age_col}{Character string for the name of the age column in \code{data}.}

\item{gender_col}{Character string for the name of the gender labels column in \code{data}.}

\item{pop_col}{Character string for the name of the projected population counts column in \code{data}.}
}
\value{
A ggplot bar charts forming a classic population pyramid.
}
\description{
Draws a population pyramid using ggplot2.
}
\examples{
library(dplyr)
library(tidyr)

# data into tidy format for plotting
df0 <- sweden1993 \%>\%
  select(age, Nx_m, Nx_f) \%>\%
  gather(key = gender, value = pop, -age) \%>\%
  mutate(gender = gsub("Nx_", "", gender))
df0

# plot
pyramid(data = df0,
        gender_lab = c("f","m"),
        age_col = "age", gender_col = "gender", pop_col = "pop")
}
