% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_open.R
\name{sp_open0}
\alias{sp_open0}
\alias{sp_open}
\title{Simple Projection Model Based on Crude Demographic Rates and Immigraiton Counts}
\usage{
sp_open0(n = NULL, N0 = NULL, b = NULL, d = NULL, e = NULL,
  I = NULL)

sp_open(n = NULL, N0 = NULL, b = NULL, d = NULL, e = NULL, I = NULL)
}
\arguments{
\item{n}{Numeric value for the number of projection periods to run the model.}

\item{N0}{Numeric value for the inital total population size.}

\item{b, d, e, I}{Numeric value for the crude birth, death and emigraiton rate and immigration count in each projection period.

If \code{sp_open0} is used a single value is required.

If \code{sp_open} is used used a vector of period specific rates is required. If a single value is be passed to \code{sp_open} a vector based on constant assumptions in all future rates (and immigration counts) will be constructed.}
}
\value{
A vector of length \code{n + 1} containing the intial total population size (\code{N0}) and the projected total population sizes given \code{b}, \code{d}, \code{e} and \code{I}.

\code{sp_open0} produces population projections based stricly on constant future rates and counts.

\code{sp_open} produces population projections based non-constant future rates and counts.
}
\description{
A simple projection model based on crude fertitly, survival and emigration rates and immigration counts,
\deqn{ N_{t+1} = \left( N_t + \frac{I_{t,t+1}}{2} \right) \times (1 + b_{t,t+1} - d_{t,t+1} - e_{t,t+1}) + \frac{I_{t,t+1}}{2} }
}
\examples{
# constant future assummptions
sp_open0(n = 20, N0 = 100, b = 0.04, d = 0.02, e = 0.01, I = 4)

# non-constant future assummptions
bb <- rnorm(n = 20, mean = 0.04, sd = 0.001)
ee <- rnorm(n = 20, mean = 0.01, sd = 0.001)
sp_open(n = 20, N0 = 100, b = bb, d = 0.02, e = ee, I = 4)
}
