% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_rate.R
\name{sp_rate}
\alias{sp_rate}
\title{Simple Projection Model with Growth Rate}
\usage{
sp_rate(n = NULL, N0 = NULL, r = NULL)
}
\arguments{
\item{n}{A numeric value for the number of projection steps to run the model.}

\item{r}{A numeric value for the assumed constant future demographic growth rate.}

\item{NO}{A numeric value for the inital total population sise.}
}
\description{
A simple projection model with a constnat growth rate for the population.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

\examples{
add_numbers(1, 2) ## returns 3

}
