% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pp.R
\name{tidy_pp}
\alias{tidy_pp}
\title{Create a Tidy Data Frame from Population Projection Matrix Objects}
\usage{
tidy_pp(proj_mat = NULL, year0 = 0, steps = NULL, age_lab = NULL,
  gender_lab = c("Female", "Male"), year_col = "Year", age_col = "Age",
  gender_col = "Gender", pop_col = "Population", denom = 1000)
}
\arguments{
\item{proj_mat}{Matrix of population projection results. Columns represent projection years, rows represent age and gender groups. If composed of both male and female groups, female population projections should be in the top half of the matrix.}

\item{year0}{Numeric value for the base year of the population projection.}

\item{steps}{Numeric value for step size of the population projection.}

\item{age_lab}{Vector containing a character string of age group labels. The length of \code{age_lab} multiplied by the length of \code{gender_lab} must be equal to the number of rows in \code{proj_mat}}

\item{gender_lab}{Vector containing a character string of gender group labels. The length of \code{age_lab} multiplied by the length of \code{gender_lab} must be equal to the number of rows in \code{proj_mat}}

\item{year_col}{Character string for the name of the year labels column.}

\item{age_col}{Character string for the name of the age labels column.}

\item{gender_col}{Character string for the name of the gender labels column.}

\item{pop_col}{Character string for the name of the projected population counts column.}

\item{denom}{Numeric value of a demominator to divide the population size of each age-gender group by.}
}
\value{
A tibble containing four columns. The first three contain information on the year, age and gender of the projection population. The fourth contain information in the projected population sizes or each year-age-gender combination.
}
\description{
Gathers a population projection matrix and converts it into a tidy data frame with labels for projection year, age and gender groups.
}
\examples{
df0 <- sweden1993

# matrix output
pp <- fccp_closed0(n = 5, x = df0$x, p = 5, Nx = df0$Nx_f,
                   sx = df0$sx_f, fx = df0$fx, sn = df0$Lx_f[1]/(5*100000),
                   tidy_output = FALSE)
pp

tidy_pp(proj_mat = pp, year0 = 1993, steps = 5,
        age_lab = df0$age, gender_lab = "Female")
}
